requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==24.1.0
    # via projecthandler (pyproject.toml)
alembic==1.17.2
    # via projecthandler (pyproject.toml)
annotated-types==0.7.0
    # via
    #   projecthandler (pyproject.toml)
    #   pydantic
anyio==4.10.0
    # via
    #   projecthandler (pyproject.toml)
    #   httpx
    #   starlette
asyncpg==0.30.0
    # via projecthandler (pyproject.toml)
bcrypt==5.0.0
    # via projecthandler (pyproject.toml)
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.2.1
    # via
    #   projecthandler (pyproject.toml)
    #   uvicorn
cryptography==46.0.3
    # via pyjwt
fastapi==0.116.1
    # via projecthandler (pyproject.toml)
greenlet==3.2.4
    # via projecthandler (pyproject.toml)
h11==0.16.0
    # via
    #   projecthandler (pyproject.toml)
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via projecthandler (pyproject.toml)
idna==3.10
    # via
    #   projecthandler (pyproject.toml)
    #   anyio
    #   httpx
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
psycopg==3.2.12
    # via projecthandler (pyproject.toml)
psycopg2-binary==2.9.10
    # via projecthandler (pyproject.toml)
pycparser==2.23
    # via cffi
pydantic==2.11.7
    # via
    #   projecthandler (pyproject.toml)
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via
    #   projecthandler (pyproject.toml)
    #   pydantic
pydantic-settings==2.10.1
    # via projecthandler (pyproject.toml)
pyjwt==2.10.1
    # via projecthandler (pyproject.toml)
python-dotenv==1.1.1
    # via
    #   projecthandler (pyproject.toml)
    #   pydantic-settings
python-multipart==0.0.20
    # via projecthandler (pyproject.toml)
sniffio==1.3.1
    # via
    #   projecthandler (pyproject.toml)
    #   anyio
sqlalchemy==2.0.43
    # via
    #   projecthandler (pyproject.toml)
    #   alembic
starlette==0.47.2
    # via
    #   projecthandler (pyproject.toml)
    #   fastapi
typing-extensions==4.14.1
    # via
    #   projecthandler (pyproject.toml)
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   projecthandler (pyproject.toml)
    #   pydantic
    #   pydantic-settings
uvicorn==0.35.0
    # via projecthandler (pyproject.toml)
